refactor(runner): split service stop logic
